requirements.txt
torch>=2.1.0
transformers>=4.41.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.11.0
trl>=0.8.0
bitsandbytes>=0.43.0
scipy
requests
tqdm
sentence-transformers
python-dotenv
faiss-cpu